Minor Update
Started logging chats so the person hosting can see what goes on in the guilds that the bot is on. Not every post on reddit has an image link, and it would show as an error in the  rich embed, so I checked if the link provided with the post is actually an image. If not, I pick another one from the list